refactor(Link/_voteForLink): Use Array#some
Instead of using Array#map, to build a list of IDs of users who voted,
and then use Array#includes to check if the user already voted, we can
use Array#some.
The main advantage is that when iterating, as soon as the callback returns
`true` the iteration stops.
See:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/some

Signed-off-by: Adrian Oprea <adrian@oprea.rocks>